Split headers and source files, replaced old particle stuff with vertex particle stuff
