+++ h/definitions.hpp
@@ -7,7 +7,7 @@
 #include <cmath>
 #include <SFML/Graphics.hpp>
 #include "utilities.hpp"
-#include "engine.hpp"
+//#include "engine.hpp"
 #include "color_gradient.hpp"
 
 using TextureMap = std::map<std::string, std::string>;
@@ -17,7 +17,7 @@ namespace Definition {
     constexpr auto offScreenLimitMult = 100.f;
 
     constexpr auto defaultMaxParticleCount = 1000;
-    const sf::Vector2f defaultParticleMotionDampening = {0.1,0.1};
+    inline sf::Vector2f defaultParticleMotionDampening = {0.1,0.1};
 
     constexpr float gravitationalConstant = 9.80665 * 1.f;
 
@@ -25,11 +25,11 @@ namespace Definition {
 
     constexpr double pi = 3.14159265358979323846;
 
-    VeX::Color_Gradient vexGradient({{102, 31, 196}, {21, 232, 255}, {255,255,255}});
+    inline VeX::Color_Gradient vexGradient({{102, 31, 196}, {21, 232, 255}, {255,255,255}});
 
     //Texture stuff
-    std::string textureDirectory = "assets/textures/";
-    TextureMap textures {
+    inline std::string textureDirectory = "assets/textures/";
+    inline TextureMap textures {
         {"defaultParticle", textureDirectory + "particle.png"},
         {"demoImage", textureDirectory + "VexedTea.png"}
     };
