+++ h/grav.hpp
@@ -8,76 +8,21 @@
 namespace VeX{
     enum class Mode{none, placing};
 
-    // struct placingModeData{
-        
-
-    // };
-
     class Grav : public State{
     private:
         Mode mode;
         Gravitational_System gravSystem;
         sf::Vector2f placingStartPoint;
     public:
-        Grav():
-            mode(Mode::none),
-            gravSystem()
-        {}
+        Grav();
         
-        void init(){
-            // gravSystem.addBody(std::make_unique<Circle_Object>(
-            //     engine->settings->getScreenCenter(),
-            //     50.f,
-            //     sf::Color(50,50,50),
-            //     // magnitude(placingStartPoint - static_cast<sf::Vector2f>(sf::Mouse::getPosition(engine->window))) * pow(10.f, 14.f),
-            //     // sf::Vector2f{0.f,0.f}
-            //     pow(10.f, 18.f),
-            //     sf::Vector2f{0.f,0.f}
-            // ));
-        }
+        void init();
         
-        void handleInput(){
-            switch(mode){
-                case Mode::none:
-                    if(engine->onLeftClick()){
-                        placingStartPoint = static_cast<sf::Vector2f>(sf::Mouse::getPosition(engine->window));
-                        mode = Mode::placing;
-                    }
-                    break;
-                case Mode::placing:
-                    if(engine->onLeftClickRelease()){
-                        gravSystem.addBody(std::make_unique<Circle_Object>(
-                            placingStartPoint,
-                            10.f,
-                            sf::Color(100,100,100),
-                            // magnitude(placingStartPoint - static_cast<sf::Vector2f>(sf::Mouse::getPosition(engine->window))) * pow(10.f, 14.f),
-                            // sf::Vector2f{0.f,0.f}
-                            pow(10.f, 15.f),
-                            placingStartPoint - static_cast<sf::Vector2f>(sf::Mouse::getPosition(engine->window))
-                        ));
-                        mode = Mode::none;
-                    }
-                    break;
-            }
-            
-            sf::Event event;
-            while (engine->window.pollEvent(event)){
-                if (event.type == sf::Event::Closed)
-                engine->window.close();
-            }
-        }
+        void handleInput();
         
-        void update(float delta){
-            gravSystem.update(delta);
-        }
+        void update(float delta);
         
-        void draw(float delta){
-            engine->window.clear(sf::Color(13, 0, 60));
-            
-            gravSystem.draw(delta);
-            
-            engine->window.display();
-        }
+        void draw(float delta);
     };
 
 } // namespace VeX
