+++ h/loading_state.hpp
@@ -1,6 +1,9 @@
 #ifndef __LOADING_STATE_HPP__
 #define __LOADING_STATE_HPP__
 
+#include <thread>
+#include <functional>
+
 #include "progress_bar.hpp"
 
 namespace VeX{
@@ -13,49 +16,15 @@ namespace VeX{
         Progress_Bar progressBar;
         
     public:
-        Loading_State(const std::string & title, float & progress, std::mutex & progressMutex, std::function<void()> progressorFunction):
-            progress(progress),
-            loadingThread(progressorFunction),
-            progressLock(progressMutex, std::defer_lock),
-            progressBar(title, {10.f,10.f}, {int(engine->settings->screenWidth)-20,100}, 10.f, Definition::vexGradient)
-        {
-            //std::cout << "Loading_State constructor" << std::endl;
-        }
+        Loading_State(const std::string & title, float & progress, std::mutex & progressMutex, std::function<void()> progressorFunction);
         
-        void init(){
-            //std::cout << "Loading_State init" << std::endl;
-        }
+        void init();
         
-        void handleInput(){
-            //Handle ur input
-            
-            sf::Event event;
-            while (engine->window.pollEvent(event)){
-                if (event.type == sf::Event::Closed)
-                engine->window.close();
-            }
-        }
+        void handleInput();
         
-        void update(float){
-            //std::cout << "Loading_State update" << std::endl;
-            progressLock.lock();
-            //std::cout << "[main thread] progress: " << progress << std::endl;
-            if(progress >= 1.f){
-                progressBar.setProgress(1.f);
-                progressLock.unlock();
-                engine->removeState();
-                loadingThread.join();
-            }else{
-                progressBar.setProgress(progress);
-                progressLock.unlock();
-            }
-        }
+        void update(float);
         
-        void draw(float){
-            progressBar.draw();
-            
-            engine->window.display();
-        }
+        void draw(float);
     };
 
 } // namespace VeX
