--- h/scene.hpp
@@ -1,18 +1,0 @@
-#ifndef __SCENE_HPP__
-#define __SCENE_HPP__
-
-namespace VeX{
-
-/// \brief
-/// The one-stop-shop for all your scene objects
-/// \details
-/// This class will hold all your objects, save them to files and load them from files and all that yadayada
-class Scene{
-private:
-
-public:
-
-};
-}
-
-#endif // __SCENE_HPP__
