+++ h/utilities.hpp
@@ -4,17 +4,18 @@
 #include <iostream>
 #include <cmath>
 #include <experimental/filesystem>
+#include "SFML/Graphics.hpp"
 namespace fs = std::experimental::filesystem;
 
 template <class T>
-class ringbuff{
+class Ringbuff{
 private:
     std::vector<T> buff;
     unsigned int size;
     unsigned int head;
     unsigned int tail;
 public:
-    explicit ringbuff(unsigned int size):
+    explicit Ringbuff(unsigned int size):
         buff(),
         size(size),
         head(0),
@@ -48,47 +49,21 @@ public:
     }
 };
 
-sf::Vector2f vector2iToVector2f(const sf::Vector2i & v){
-    return {static_cast<float>(v.x), static_cast<float>(v.y)};
-}
+sf::Vector2f vector2iToVector2f(const sf::Vector2i & v);
 
-float angleRadians(const sf::Vector2f & v){
-    float angle = atan2(v.y, v.x)/* *180/3.14159265358979323846 */;
-    // if(angle < 0){
-    //     angle += 360;
-    // }
-    return angle;
-}
+float angleRadians(const sf::Vector2f & v);
 
-sf::Vector2f pow(const sf::Vector2f & v, int power){
-    return sf::Vector2f(pow(v.x, power), pow(v.y, power));
-}
+sf::Vector2f pow(const sf::Vector2f & v, int power);
 
-float distance(const sf::Vector2f & point1, const sf::Vector2f & point2){
-    return sqrt(pow(point1.x - point2.x, 2) + pow(point1.y - point2.y, 2));
-}
+float distance(const sf::Vector2f & point1, const sf::Vector2f & point2);
 
-float magnitude(const sf::Vector2f & v){
-    return distance(v, {0.f,0.f});
-}
+float magnitude(const sf::Vector2f & v);
 
-sf::Vector2f normalize(const sf::Vector2f & v){
-    return v / magnitude(v);
-}
+sf::Vector2f normalize(const sf::Vector2f & v);
 
-unsigned int countFilesInFolder(const std::string & folderPath){
-    unsigned int count = 0;
-    for(const auto & entry : fs::directory_iterator(folderPath)){
-        if(fs::is_regular_file(entry)){
-            count++;
-        }
-    }
-    return count;
-}
+unsigned int countFilesInFolder(const std::string & folderPath);
 
-sf::Vector2f getPositionOnCircle(const sf::Vector2f & centerPos, float radius, float angleDegrees){
-    return sf::Vector2f(centerPos.x + (radius*cos(angleDegrees*(3.14159265358979323846/180.f))), centerPos.y + (radius*sin(angleDegrees*(3.14159265358979323846/180.f))));
-}
+sf::Vector2f getPositionOnCircle(const sf::Vector2f & centerPos, float radius, float angleDegrees);
 
 template<typename T>
 inline sf::Vector2f operator/(
@@ -213,19 +188,10 @@ inline sf::Vector2f operator/(
     return {left.x / right, left.y / right};
 }
 
-std::ostream& operator<<(std::ostream &os, const sf::Vector2f & vec){
-    os << "(" << vec.x <<", "<< vec.y << ")";
-    return os;
-}
+std::ostream& operator<<(std::ostream &os, const sf::Vector2f & vec);
 
-std::ostream& operator<<(std::ostream &os, const sf::Vector2i & vec){
-    os << "(" << vec.x <<", "<< vec.y << ")";
-    return os;
-}
+std::ostream& operator<<(std::ostream &os, const sf::Vector2i & vec);
 
-std::ostream& operator<<(std::ostream &os, const sf::Vector2u & vec){
-    os << "(" << vec.x <<", "<< vec.y << ")";
-    return os;
-}
+std::ostream& operator<<(std::ostream &os, const sf::Vector2u & vec);
 
 #endif // __VECTOR_OPERATORS_HPP__
